moved entry point back to __main__, added vision support to AI Clients
