--- entry_point.py
@@ -1,6 +1,0 @@
-"""Entry point for PyInstaller to create an executable"""
-
-from llmcord.main import main
-
-if __name__ == "__main__":
-    main()
